requirements.txt
bleach==3.3.1
build==0.5.1
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
cryptography==3.4.7
docutils==0.17.1
google-api-core==1.30.0
google-auth==1.32.1
google-cloud-bigquery==2.20.0
google-cloud-core==1.7.1
google-crc32c==1.1.2
google-resumable-media==1.3.1
googleapis-common-protos==1.53.0
grpcio==1.38.1
idna==2.10
importlib-metadata==4.6.1
jeepney==0.7.0
Jinja2==3.0.1
keyring==23.0.1
MarkupSafe==2.0.1
packaging==20.9
pep517==0.11.0
pkginfo==1.7.1
proto-plus==1.19.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.9.0
pyparsing==2.4.7
python-bigquery-validator==0.0.2a6
pytz==2021.1
readme-renderer==29.0
requests==2.25.1
requests-toolbelt==0.9.1
rfc3986==1.5.0
rsa==4.7.2
SecretStorage==3.3.1
six==1.16.0
toml==0.10.2
tomli==1.0.4
tqdm==4.61.2
twine==3.4.2
typing-extensions==3.10.0.0
urllib3==1.26.6
webencodings==0.5.1
zipp==3.5.0
